+++ sites/geohub/src/components/Settings.svelte
@@ -1,74 +1,183 @@
 <script lang="ts">
   import { enhance } from '$app/forms'
+  import { base } from '$app/paths'
+  import { invalidateAll, afterNavigate, goto } from '$app/navigation'
   import {
     DatasetSortingColumns,
     MapSortingColumns,
     NumberOfClassesMaximum,
     NumberOfClassesMinimum,
     LimitOptions,
+    RasterResamplingMethods,
+    IconOverlapPriority,
   } from '$lib/config/AppConfig'
   import { toast } from '@zerodevx/svelte-toast'
   import { page } from '$app/stores'
   import { DefaultUserConfig, type UserConfig } from '$lib/config/DefaultUserConfig'
-  import { invalidateAll } from '$app/navigation'
   import type { SidebarPosition } from '$lib/types'
   import RangeSlider from 'svelte-range-slider-pips'
+  import { Radios } from '@undp-data/svelte-undp-design'
   import FieldControl from './controls/FieldControl.svelte'
+  import { ClassificationMethods } from '$lib/config/AppConfig'
+  import { initTippy } from '$lib/helper'
+  import { clean } from '$lib/helper/index.js'
+  import { spriteImageList } from '$stores'
+  import IconImagePickerCard from '$components/controls/vector-styles/IconImagePickerCard.svelte'
+  import { LineTypes } from '$lib/config/AppConfig/LineTypes'
 
+  // preserve previous page URL
+  let previousPage: string = base
+  afterNavigate(({ from }) => {
+    if (from?.url) {
+      previousPage = `${from?.url.pathname}${from?.url.search}`
+    }
+  })
+
+  const tippy = initTippy()
   let userSettings: UserConfig = $page.data.config
   let isSubmitting = false
-  let sideBarPosition: SidebarPosition = userSettings.SidebarPosition || 'left'
+  let sideBarPosition: SidebarPosition = userSettings.SidebarPosition
   let lineWidth = [userSettings.LineWidth]
   let numberOfClasses = [userSettings.NumberOfClasses]
-  let isExpanded = true
-  let activeSettingTab = 'Map'
+  let labelFontSize = [userSettings.LabelFontSize]
+  let labelHaloWidth = [userSettings.LabelHaloWidth]
+  let iconSize = [userSettings.IconSize]
+  let layerOpacity = [userSettings.LayerOpacity]
+  let selectedIcon = userSettings.IconImage
+
+  let linePattern = LineTypes.find((t) => t.title === userSettings.LinePattern)?.title
+  const setLinePatterns = () => {
+    const pattern = LineTypes.map((type) => {
+      const label = `
+          ${type.title}
+          <span
+            style="font-family: monospace;position:relative;left: 10px;top:-4px;position:relative;font-weight: bold;">
+            ${type.pattern}
+          </span>`
+
+      return {
+        label: label,
+        value: type.title,
+      }
+    })
+    return pattern
+  }
+
+  let linePatterns = setLinePatterns()
+
+  $: iconImageSrc = $spriteImageList.find((i) => i.alt === selectedIcon)?.src
+  let tooltipContent: HTMLElement
+
+  let settingTabs = [
+    {
+      title: 'GeoHub Home',
+      hash: 'Home',
+      icon: 'fa-solid fa-home',
+      subSettings: [
+        { title: 'Main', hash: 'main' },
+        { title: 'Legend', hash: 'legend' },
+        { title: 'Line', hash: 'line' },
+        { title: 'Point', hash: 'point' },
+        // { title: 'Polygon', hash: 'polygon' },
+        { title: 'Raster', hash: 'raster' },
+        { title: 'Label', hash: 'label' },
+      ],
+    },
+    {
+      title: 'My data',
+      hash: 'data',
+      icon: 'fa-solid fa-server',
+    },
+    {
+      title: 'Maps',
+      hash: 'maps',
+      icon: 'fa-solid fa-map',
+    },
+  ]
+  const hash = $page.url.hash
+  let activeTab = settingTabs[0].subSettings ? settingTabs[0].subSettings[0] : settingTabs[0]
+
+  if (hash) {
+    let tab = settingTabs.find((t) => `#${t.hash}` === hash)
+    if (tab) {
+      activeTab = tab
+    }
+  }
+  let activeSettingTab = activeTab.title
 
   const DatasetLimitOptions = LimitOptions.includes(DefaultUserConfig.DatasetSearchLimit)
     ? LimitOptions
     : [...LimitOptions, DefaultUserConfig.DatasetSearchLimit].sort((a, b) => a - b)
 
-  export let headerHeight: number
+  const resetToDefault = () => {
+    userSettings = JSON.parse(JSON.stringify(DefaultUserConfig))
+    sideBarPosition = userSettings.SidebarPosition
+    lineWidth = [userSettings.LineWidth]
+    numberOfClasses = [userSettings.NumberOfClasses]
+    labelFontSize = [userSettings.LabelFontSize]
+    labelHaloWidth = [userSettings.LabelHaloWidth]
+    iconSize = [userSettings.IconSize]
+    layerOpacity = [userSettings.LayerOpacity]
+    linePattern = LineTypes.find((t) => t.title === userSettings.LinePattern)?.title
+    linePatterns = setLinePatterns()
+    selectedIcon = userSettings.IconImage
+    toast.push('Settings were reset. Please click apply button to save them.')
+  }
 
-  const collapseMiniMenu = () => {
-    isExpanded = !isExpanded
+  const backToPreviousPage = () => {
+    goto(previousPage, {
+      invalidateAll: true,
+    })
   }
 </script>
 
-<div
-  class="columns is-one-quarter ml-auto mr-auto settings-page"
-  style="margin-top: {headerHeight}px;">
+<div class="columns is-one-quarter ml-auto mr-auto settings-page">
   <div class="column is-2">
     <aside class="menu">
+      <p class="menu-label">Settings</p>
       <ul class="menu-list">
-        <li>
-          <a
-            on:click={collapseMiniMenu}
-            href="#">GeoHub Settings</a>
-          <ul style="display: {!isExpanded ? 'none' : ''}">
+        {#each settingTabs as tab}
+          {#if tab.subSettings}
             <li>
               <a
-                class={activeSettingTab === 'Map' ? 'selected' : ''}
-                on:click={() => (activeSettingTab = 'Map')}
-                href="#">Map Settings</a>
+                class={activeSettingTab === tab.title ? 'is-active' : ''}
+                href="#{tab.hash}">
+                <span class="icon">
+                  <i class="{tab.icon} {activeSettingTab === tab.title ? 'has-text-white' : 'has-text-link'}" />
+                </span>
+                {tab.title}
+              </a>
+              <ul>
+                {#each tab.subSettings as subSetting}
+                  <li>
+                    <a
+                      class={activeSettingTab === subSetting.title ? 'is-active' : ''}
+                      on:click={() => (activeSettingTab = subSetting.title)}
+                      href="#{subSetting.title}">
+                      {subSetting.title}
+                    </a>
+                  </li>
+                {/each}
+              </ul>
             </li>
+          {:else}
             <li>
               <a
-                class={activeSettingTab === 'Search' ? 'selected' : ''}
-                on:click={() => (activeSettingTab = 'Search')}
-                href="#">Search Settings</a>
+                class={activeSettingTab === tab.title ? 'is-active' : ''}
+                on:click={() => (activeSettingTab = tab.title)}
+                href="#{tab.hash}">
+                <span class="icon">
+                  <i class="{tab.icon} {activeSettingTab === tab.title ? 'has-text-white' : 'has-text-link'}" />
+                </span>
+                {tab.title}
+              </a>
             </li>
-            <li>
-              <a
-                class={activeSettingTab === 'Legend' ? 'selected' : ''}
-                on:click={() => (activeSettingTab = 'Legend')}
-                href="#">Legend Settings</a>
-            </li>
-          </ul>
-        </li>
+          {/if}
+        {/each}
       </ul>
     </aside>
   </div>
-  <div class="column is-two-fifths m-auto">
+  <div class="column is-three-fifths m-auto">
     <form
       action="?/save"
       method="post"
@@ -84,18 +193,25 @@
           isSubmitting = false
         }
       }}>
-      <section
-        class="content {activeSettingTab !== 'Map' ? 'is-hidden' : ''}"
-        id="Geohub">
-        <h1 class="title">Map Settings</h1>
+      {#if previousPage}
+        <button
+          type="button"
+          disabled={isSubmitting}
+          class="button is-link"
+          on:click={backToPreviousPage}>
+          Back to previous page
+        </button>
+      {/if}
 
+      <section class="content {activeSettingTab !== 'Main' ? 'is-hidden' : ''}">
+        <p class="title is-4">Layout Settings</p>
         <FieldControl title="Sidebar Position">
           <div slot="help">Select sidebar position of main GeoHub page.</div>
           <div slot="control">
-            <div class="columns">
+            <div class="columns is-mobile">
               <label class="column">
                 <input
-                  on:select={() => userSettings.SidebarPosition === 'left'}
+                  on:select={() => sideBarPosition === 'left'}
                   type="radio"
                   name="SidebarPosition"
                   value="left"
@@ -107,7 +223,7 @@
               </label>
               <label class="column">
                 <input
-                  on:select={() => userSettings.SidebarPosition === 'right'}
+                  on:select={() => sideBarPosition === 'right'}
                   type="radio"
                   name="SidebarPosition"
                   value="right"
@@ -120,28 +236,8 @@
             </div>
           </div>
         </FieldControl>
-      </section>
-      <section
-        class="content {activeSettingTab !== 'Search' ? 'is-hidden' : ''}"
-        id="Search">
-        <h1 class="title">Search Settings</h1>
-
-        <FieldControl title="Default search Limit in data and maps page">
-          <div slot="help">The number of items to search at data page and maps page</div>
-          <div slot="control">
-            <div class="select is-fullwidth">
-              <select
-                name="SearchLimit"
-                bind:value={userSettings.SearchLimit}>
-                {#each LimitOptions as limit}
-                  <option value={limit}>{limit}</option>
-                {/each}
-              </select>
-            </div>
-          </div>
-        </FieldControl>
-
-        <FieldControl title="Default search Limit in main GeoHub page">
+        <p class="title is-4">Search Settings</p>
+        <FieldControl title="Default search Limit">
           <div slot="help">The number of items to search at data tab in main GeoHub page.</div>
           <div slot="control">
             <div class="select is-fullwidth">
@@ -156,7 +252,7 @@
           </div>
         </FieldControl>
 
-        <FieldControl title="Default search query operator in main GeoHub page">
+        <FieldControl title="Default search query operator">
           <div slot="help">
             Change searching operator to either 'AND' or 'OR'. 'AND' enables you to search datasets which exactly match
             all keyword. 'OR' allows you to search wider range of results by matching at least a word.
@@ -180,7 +276,7 @@
           </div>
         </FieldControl>
 
-        <FieldControl title="Default sort setting in main GeoHub page">
+        <FieldControl title="Default sort setting">
           <div slot="help">Change sort setting for the search result on datasets.</div>
           <div slot="control">
             <div class="select is-fullwidth">
@@ -197,7 +293,7 @@
           </div>
         </FieldControl>
 
-        <FieldControl title="Defaut tag search operator in main GeoHub page & data page">
+        <FieldControl title="Defaut tag search operator">
           <div slot="help">
             Change searching operator for tag filter to either 'AND' or 'OR'. 'AND' enables you to search datasets which
             exactly match all tags you selected. 'OR' allows you to search wider range of results by matching at least a
@@ -221,43 +317,27 @@
             </div>
           </div>
         </FieldControl>
-
-        <FieldControl title="Default sort setting in Data page">
-          <div slot="help">Change sort setting for the search result on datasets.</div>
-          <div slot="control">
-            <div class="select is-fullwidth">
-              <select
-                name="DataPageSortingColumn"
-                bind:value={userSettings.DataPageSortingColumn}>
-                {#each DatasetSortingColumns as column}
-                  <option value={column.value}>{column.label}</option>
-                {/each}
-              </select>
-            </div>
-          </div>
-        </FieldControl>
-
-        <FieldControl title="Default sort setting in Maps page">
-          <div slot="help">Change sort setting for the search result on datasets.</div>
-          <div slot="control">
-            <div class="select is-fullwidth">
-              <select
-                name="MapPageSortingColumn"
-                bind:value={userSettings.MapPageSortingColumn}>
-                {#each MapSortingColumns as column}
-                  <option value={column.value}>{column.label}</option>
-                {/each}
-              </select>
-            </div>
-          </div>
-        </FieldControl>
       </section>
-
       <section class="content {activeSettingTab !== 'Legend' ? 'is-hidden' : ''}">
-        <h1 class="title">Legend Settings</h1>
-
+        <p class="title is-4">Legend Settings</p>
+        <FieldControl title="Default Classification Method">
+          <div slot="help">Change the default classification method</div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="ClassificationMethod"
+                bind:value={userSettings.ClassificationMethod}>
+                {#each ClassificationMethods as classificationMethod}
+                  <option value={classificationMethod.value}>{classificationMethod.label}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
         <FieldControl title="Default number of classes">
-          <div slot="help">The default number of classes in classify legend for vector layer and raster layer</div>
+          <div slot="help">
+            Change the default number of classes in classify legend for vector layer and raster layer
+          </div>
           <div slot="control">
             <div class="control">
               <RangeSlider
@@ -267,18 +347,47 @@
                 max={NumberOfClassesMaximum}
                 step={1}
                 pips
+                springValues={{
+                  stiffness: 1,
+                  damping: 1,
+                }}
                 first="label"
                 last="label"
                 rest={false} />
+              <input
+                type="hidden"
+                name="NumberOfClasses"
+                bind:value={numberOfClasses[0]} />
             </div>
           </div>
         </FieldControl>
-
-        <input
-          type="hidden"
-          name="NumberOfClasses"
-          bind:value={numberOfClasses[0]} />
-
+        <FieldControl title="Default Layer Opacity">
+          <div slot="help">Change Default Layer Opacity</div>
+          <div slot="control">
+            <RangeSlider
+              bind:values={layerOpacity}
+              float
+              min={0}
+              max={100}
+              step={1}
+              pips
+              first="label"
+              last="label"
+              springValues={{
+                stiffness: 1,
+                damping: 1,
+              }}
+              suffix="%"
+              rest={false} />
+            <input
+              type="hidden"
+              bind:value={layerOpacity[0]}
+              name="LayerOpacity" />
+          </div>
+        </FieldControl>
+      </section>
+      <section class="content {activeSettingTab !== 'Line' ? 'is-hidden' : ''}">
+        <p class="title is-4">Line Visualization Settings</p>
         <FieldControl title="Default line width">
           <div slot="help">
             The default line width in <b>line</b> vector layer legend tab.
@@ -299,22 +408,359 @@
                 first="label"
                 last="label"
                 rest={false} />
+              <input
+                type="hidden"
+                name="LineWidth"
+                bind:value={lineWidth[0]} />
+            </div>
+          </div>
+        </FieldControl>
+        <FieldControl title="Default line pattern">
+          <div slot="help">
+            The default pattern in <b>line</b> vector layer legend tab.
+          </div>
+          <div slot="control">
+            <div
+              class="line-pattern-view-container"
+              data-testid="line-pattern-view-container">
+              <Radios
+                groupName="LinePattern"
+                bind:radios={linePatterns}
+                bind:value={linePattern}
+                allowHtml={true}
+                isVertical={true} />
+            </div>
+          </div>
+        </FieldControl>
+      </section>
+      <section class="content {activeSettingTab !== 'Point' ? 'is-hidden' : ''}">
+        <p class="title is-4">Point Visualization Settings</p>
+        <FieldControl
+          title="Icon Symbol"
+          class="icon-selector">
+          <div slot="help">Pick the default icon symbol for symbol layers</div>
+          <div slot="control">
+            <div
+              style="cursor: pointer"
+              use:tippy={{ content: tooltipContent }}
+              class="card"
+              data-testid="icon-image-picker-card-container">
+              <div class="card-content">
+                <div class="media is-flex is-justify-content-center">
+                  <figure
+                    class={`image is-24x24`}
+                    data-testid="icon-figure">
+                    <img
+                      data-testid="icon-image"
+                      type="image"
+                      src={iconImageSrc}
+                      alt={clean(selectedIcon)}
+                      title={clean(selectedIcon)}
+                      style="width:24px; height:24px; color: white;" />
+                  </figure>
+                </div>
+                <div
+                  class="content is-size-7 columns is-gapless"
+                  style="padding-top: 5px;">
+                  <div
+                    class="column is-flex is-justify-content-center sprite-image-title"
+                    title={selectedIcon}>
+                    {clean(selectedIcon)}
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div
+              style="max-height: 400px; overflow-y: auto; overflow-x: hidden; max-width: fit-content"
+              class="tooltip"
+              data-testid="tooltip"
+              bind:this={tooltipContent}>
+              <div class="columns m-2 is-multiline is-justify-content-space-evenly">
+                {#each $spriteImageList as image}
+                  <IconImagePickerCard
+                    on:iconSelected={(e) => (selectedIcon = e.detail.iconImageAlt)}
+                    iconImageAlt={image.alt}
+                    iconImageSrc={image.src}
+                    withinForm={true}
+                    isSelected={selectedIcon === image.alt ? true : false} />
+                {/each}
+              </div>
+            </div>
+            <input
+              type="hidden"
+              value={selectedIcon}
+              name="IconImage" />
+          </div>
+        </FieldControl>
+        <FieldControl title="Icon Overlap Priority">
+          <div slot="help">
+            Change Icon Overlap priority. When <b>Never</b> is selected, the icon will be hidden if it collides with any
+            other previously drawn symbol. When <b>Always</b> is selected, the icon will be visible even if it collides
+            with any other previously drawn symbol. When <b>Cooperative</b> is selected, If the icon collides with another
+            previously drawn symbol, the overlap mode for that symbol is checked. If the previous symbol was placed using
+            never overlap mode, the new icon is hidden. If the previous symbol was placed using always or cooperative overlap
+            mode, the new icon is visible.
+          </div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="IconOverlapPriority"
+                bind:value={userSettings.IconOverlapPriority}>
+                {#each IconOverlapPriority as overlapPriority}
+                  <option value={overlapPriority.value}>{overlapPriority.label}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+        <FieldControl title="Icon Size">
+          <div slot="help">Change icon size for symbol layers</div>
+          <div slot="control">
+            <RangeSlider
+              bind:values={iconSize}
+              float
+              min={0}
+              max={5}
+              step={0.1}
+              pips
+              first="label"
+              last="label"
+              springValues={{
+                stiffness: 1,
+                damping: 1,
+              }}
+              rest={false} />
+            <input
+              type="hidden"
+              bind:value={iconSize[0]}
+              name="IconSize" />
+          </div>
+        </FieldControl>
+      </section>
+      <!--      <section class="content {activeSettingTab !== 'Polygon' ? 'is-hidden' : ''}" style="height:400px">-->
+      <!--        <p class="title is-4">Polygon Visualization Settings</p>-->
+      <!--        <FieldControl title="Default polygon fill color">-->
+      <!--          <div slot="help">Change default polygon fill color</div>-->
+      <!--          <div slot="control">-->
+      <!--            <div class="field has-addons">-->
+      <!--              <div class="control is-expanded">-->
+      <!--                <input-->
+      <!--                  type="color"-->
+      <!--                  class="input"-->
+      <!--                  name="PolygonFillColor"-->
+      <!--                  bind:value={userSettings.PolygonFillColor}-->
+      <!--                  data-testid="polygon-fill-color-input" />-->
+      <!--              </div>-->
+      <!--              <div class="control">-->
+      <!--                <div class="button is-static">-->
+      <!--                  <i class="fas fa-palette" />-->
+      <!--                </div>-->
+      <!--              </div>-->
+      <!--            </div>-->
+      <!--          </div>-->
+      <!--        </FieldControl>-->
+      <!--      </section>-->
+      <section class="content {activeSettingTab !== 'Raster' ? 'is-hidden' : ''}">
+        <p class="title is-4">Raster Visualization Settings</p>
+        <FieldControl title="Default raster resampling method">
+          <div slot="help">
+            Change raster resampling method
+            <p>
+              <b>Bili-near</b> filtering interpolates pixel values using the weighted average of the four closest original
+              source pixels creating a smooth but blurry look when overscaled
+            </p>
+            <p>
+              <b>Nearest neighbour</b> filtering interpolates pixel values using the weighted average of the four closest
+              original source pixels creating a smooth but blurry look when overscaled
+            </p>
+          </div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="RasterResamplingMethod"
+                bind:value={userSettings.RasterResamplingMethod}>
+                {#each RasterResamplingMethods as resamplingMethod}
+                  <option value={resamplingMethod.value}>{resamplingMethod.label}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+      </section>
+      <section class="content {activeSettingTab !== 'Label' ? 'is-hidden' : ''}">
+        <p class="title is-4">Label Settings</p>
+        <FieldControl title="Default label font size">
+          <div slot="help">Change default label font size</div>
+          <div slot="control">
+            <RangeSlider
+              bind:values={labelFontSize}
+              float
+              min={0}
+              max={32}
+              step={0.5}
+              pips
+              first="label"
+              last="label"
+              springValues={{
+                stiffness: 1,
+                damping: 1,
+              }}
+              rest={false} />
+            <input
+              type="hidden"
+              bind:value={labelFontSize[0]}
+              name="LabelFontSize" />
+          </div>
+        </FieldControl>
+
+        <FieldControl title="Default label halo width">
+          <div slot="help">Change default halo size for labels.</div>
+          <div slot="control">
+            <RangeSlider
+              bind:values={labelHaloWidth}
+              float
+              min={0}
+              max={10}
+              step={0.1}
+              pips
+              first="label"
+              last="label"
+              springValues={{
+                stiffness: 1,
+                damping: 1,
+              }}
+              rest={false} />
+            <input
+              type="hidden"
+              bind:value={labelHaloWidth[0]}
+              name="LabelHaloWidth" />
+          </div>
+        </FieldControl>
+      </section>
+      <section class="content {activeSettingTab !== settingTabs[1].title ? 'is-hidden' : ''}">
+        <p class="title is-4">Search Settings</p>
+        <FieldControl title="Default search Limit">
+          <div slot="help">The number of items to search at data page and maps page</div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="SearchLimit"
+                bind:value={userSettings.SearchLimit}>
+                {#each LimitOptions as limit}
+                  <option value={limit}>{limit}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+        <FieldControl title="Default search query operator">
+          <div slot="help">
+            Change searching operator to either 'AND' or 'OR'. 'AND' enables you to search datasets which exactly match
+            all keyword. 'OR' allows you to search wider range of results by matching at least a word.
+          </div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="DatasetSearchQueryOperator"
+                bind:value={userSettings.DatasetSearchQueryOperator}>
+                {#each ['and', 'or'] as operator}
+                  <option value={operator}>
+                    {#if operator === 'and'}
+                      Match all words typed (AND)
+                    {:else}
+                      Match at least a word typed (OR)
+                    {/if}
+                  </option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+        <FieldControl title="Defaut tag search operator">
+          <div slot="help">
+            Change searching operator for tag filter to either 'AND' or 'OR'. 'AND' enables you to search datasets which
+            exactly match all tags you selected. 'OR' allows you to search wider range of results by matching at least a
+            tag selected.
+          </div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="TagSearchOperator"
+                bind:value={userSettings.TagSearchOperator}>
+                {#each ['and', 'or'] as operator}
+                  <option value={operator}>
+                    {#if operator === 'and'}
+                      Match all selected tags (AND)
+                    {:else}
+                      Match at least a tag selected (OR)
+                    {/if}
+                  </option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+        <FieldControl title="Default sort setting">
+          <div slot="help">Change sort setting for the search result on datasets.</div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="DataPageSortingColumn"
+                bind:value={userSettings.DataPageSortingColumn}>
+                {#each DatasetSortingColumns as column}
+                  <option value={column.value}>{column.label}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
+      </section>
+      <section class="content {activeSettingTab !== settingTabs[2].title ? 'is-hidden' : ''}">
+        <p class="title is-4">Search Settings</p>
+        <FieldControl title="Default search Limit">
+          <div slot="help">The number of items to search at data page and maps page</div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="SearchLimit"
+                bind:value={userSettings.SearchLimit}>
+                {#each LimitOptions as limit}
+                  <option value={limit}>{limit}</option>
+                {/each}
+              </select>
             </div>
           </div>
         </FieldControl>
 
-        <input
-          type="hidden"
-          name="LineWidth"
-          bind:value={lineWidth[0]} />
+        <FieldControl title="Default sort setting">
+          <div slot="help">Change sort setting for the search result on datasets.</div>
+          <div slot="control">
+            <div class="select is-fullwidth">
+              <select
+                name="MapPageSortingColumn"
+                bind:value={userSettings.MapPageSortingColumn}>
+                {#each MapSortingColumns as column}
+                  <option value={column.value}>{column.label}</option>
+                {/each}
+              </select>
+            </div>
+          </div>
+        </FieldControl>
       </section>
       <div class="field is-grouped is-grouped-centered">
         <div class="control">
           <button
+            type="button"
+            disabled={isSubmitting}
+            class="button is-link"
+            on:click={resetToDefault}>
+            Reset to default
+          </button>
+          <button
             formaction="?/save"
             type="submit"
             class="button is-primary {isSubmitting ? 'is-loading' : ''}">
-            Submit
+            Apply
           </button>
         </div>
       </div>
@@ -328,28 +774,6 @@
   }
   .content.is-hidden:not(:first-of-type) {
     display: none;
-  }
-  .menu-list li {
-    margin-bottom: 0.5rem;
-  }
-  .menu-list a.selected {
-    background-color: #3273dc;
-    color: white;
-  }
-  .menu-list a {
-    cursor: pointer;
-  }
-
-  .card {
-    z-index: -1;
-  }
-
-  .sidebar-col:hover {
-    cursor: pointer;
-  }
-
-  .selected-sidebar {
-    border: 2px solid #3273dc;
   }
 
   [type='radio'] {
@@ -369,5 +793,10 @@
 
   .sidebar-image {
     box-shadow: #0a0a0a 0 0 2px 0;
+  }
+
+  .icon-selector {
+    display: flex;
+    flex-direction: column;
   }
 </style>