+++ sites/geohub/src/components/UserAccount.svelte
@@ -4,6 +4,7 @@
   import { Button } from '@undp-data/svelte-undp-design'
   import chroma from 'chroma-js'
   import { goto } from '$app/navigation'
+  import { clickOutside } from 'svelte-use-click-outside'
 
   let panelWidth = '350px'
   let dropdownActive = false
@@ -41,7 +42,7 @@
             class="avatar" />
         {:else}
           <span
-            class="initial-avator is-flex is-justify-content-center	is-align-items-center"
+            class="initial-avator is-flex is-justify-content-center is-align-items-center"
             style="background-color: {chroma.random()}">
             {#each names as name}
               <p
@@ -57,7 +58,10 @@
     <div
       class="dropdown-menu"
       style="max-width: {panelWidth}"
-      role="menu">
+      role="menu"
+      use:clickOutside={() => {
+        dropdownActive = false
+      }}>
       <div class="dropdown-content">
         <div class="dropdown-item has-text-centered">
           <p class="title mb-2 is-4">{$page.data.session.user.name}</p>