+++ sites/geohub/src/components/controls/VectorParamsPanel.svelte
@@ -3,7 +3,14 @@
   import RangeSlider from 'svelte-range-slider-pips'
   import { Loader } from '@undp-data/svelte-undp-design'
 
-  import { fetchUrl, getLayerSourceUrl, getLayerStyle, loadMap, updateLayerURL } from '$lib/helper'
+  import {
+    fetchUrl,
+    getLayerSourceUrl,
+    getLayerStyle,
+    loadArgumentsInDynamicLayers,
+    loadMap,
+    updateLayerURL,
+  } from '$lib/helper'
   import { map, layerList } from '$stores'
   /*EXPORTS*/
   export let layerId
@@ -31,17 +38,12 @@
 
   const init = async () => {
     const isLoaded = await loadMap($map)
-    const metaUrl = layerUrl.replace('/{z}/{x}/{y}.pbf', '.json')
-    const jsonString = await fetchUrl(metaUrl)
-    args = JSON.parse(jsonString.arguments[0].default)
-
+    args = await loadArgumentsInDynamicLayers(layerUrl)
     for (const [k, v] of Object.entries(args)) {
       defaultArgs[k] = { value: Number(v.value) }
     }
     selectedArgs = getArgumentsInURL() || selectedArgs
     if (selectedArgs) currentSelectedArg = Object.keys(selectedArgs).at(-1)
-    //console.log('hinit', JSON.stringify(selectedArgs, null, 2))
-
     return isLoaded
   }
 
@@ -130,7 +132,7 @@
             currentSelectedArg = argId
           }}>
           <div
-            class="is-flex is-flex-direction-row is-flex-wrap-nowrap is-justify-content-space-evenly "
+            class="is-flex is-flex-direction-row is-flex-wrap-nowrap is-justify-content-space-evenly"
             style="height:80px">
             <div class="has-text-primary">
               <i
@@ -141,7 +143,7 @@
           </div>
         </div>
         <footer class="card-footer">
-          <div class="content m-auto has-text-primary  has-text-weight-bold">
+          <div class="content m-auto has-text-primary has-text-weight-bold">
             {argId in selectedArgs ? selectedArgs[argId].value : value}
             {units}
           </div>