+++ sites/geohub/src/components/data-view/DataCard.svelte
@@ -26,6 +26,8 @@
   export let isStarOnly = false
 
   let defaultLineWidth = $page.data.config.LineWidth
+  let layerOpacity = $page.data.config.LayerOpacity / 100
+
   let nodeRef: HTMLElement
   let defaultColor: string = undefined
   let defaultColormap: string = undefined
@@ -82,7 +84,7 @@
         } else {
           // COG
           const rasterInfo = metadata as RasterTileMetadata
-          const rasterTile = new RasterTileData(titilerUrl, feature, rasterInfo)
+          const rasterTile = new RasterTileData(titilerUrl, feature, rasterInfo, layerOpacity)
           const data = await rasterTile.add($map, defaultColormap)
           $layerList = [
             {