+++ sites/geohub/src/lib/config/DefaultUserConfig/ClassificationMethod.ts
@@ -1,0 +1,3 @@
+import { ClassificationMethodTypes } from '$lib/config/AppConfig'
+
+export const ClassificationMethod: ClassificationMethodTypes = ClassificationMethodTypes.EQUIDISTANT