+++ sites/geohub/src/lib/config/DefaultUserConfig/LinePattern.ts
@@ -1,0 +1,1 @@
+export const LinePattern = 'solid'