+++ packages/current-location/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @undp-data/current-location
 
+## 0.1.3
+
+### Patch Changes
+
+- c8f5b2f1: chore: updated all npm packages except maplibre-gl
+- c8f5b2f1: fix: bug of setting feature state when the zoom level is 5 or above in current-location plugin
+
 ## 0.1.2
 
 ### Patch Changes